fix(ci): fix Windows build errors in CLI and llama32_demo
- Silence "unused parameter" warnings/errors for `cwd` and
  `env_overrides` in `tooling/cli/main.cpp` when compiling for Windows.
- Fix "conversion from 'size_t' to 'int', possible loss of data" error
  in `examples/llama32_demo.cpp` by explicitly using `static_cast<int>`.

+++ examples/llama32_demo.cpp
@@ -137,7 +137,7 @@ int main() {
 
   // Add hash to symbol pool
   weight_names.push_back(hash_str);
-  int hash_symbol_index = weight_names.size();  // 1-based index (last element)
+  int hash_symbol_index = static_cast<int>(weight_names.size());  // 1-based index (last element)
 
   program.symbol_pool = weight_names;
